fix(review): address PR #1070 review feedback
Address review comments from @MRashad26:

- scheduled-send-tooltip.pipe.ts: add pure scheduledSendTooltip and
  scheduledSendAriaLabel pipes to replace component method calls in
  template bindings (per @MRashad26)
- surveys-table.component.html: use pipes for tooltip, tooltipDisabled,
  tabindex, and aria-label bindings instead of @let + component methods
  (per @MRashad26)
- surveys-table.component.ts: remove getScheduledSendTooltip,
  getScheduledStatusAriaLabel, and getStatusLabel helpers now handled
  by pipes (per @MRashad26)

Resolves 2 review threads.
