+++ apps/lfx-one/src/app/modules/surveys/components/surveys-table/surveys-table.component.ts
@@ -13,11 +13,12 @@ import { InputTextComponent } from '@components/input-text/input-text.component'
 import { SelectComponent } from '@components/select/select.component';
 import { TableComponent } from '@components/table/table.component';
 import { TagComponent } from '@components/tag/tag.component';
-import { SURVEY_LABEL, SURVEY_STATUS_LABELS, SURVEY_TYPE_LABELS, SurveyStatus } from '@lfx-one/shared';
+import { SURVEY_LABEL, SURVEY_TYPE_LABELS, SurveyStatus } from '@lfx-one/shared';
 import { FilterPillOption, Survey } from '@lfx-one/shared/interfaces';
 import { getSurveyDisplayStatus } from '@lfx-one/shared/utils';
 import { DueDateLabelColorPipe } from '@pipes/due-date-label-color.pipe';
 import { DueDateLabelPipe } from '@pipes/due-date-label.pipe';
+import { ScheduledSendAriaLabelPipe, ScheduledSendTooltipPipe } from '@pipes/scheduled-send-tooltip.pipe';
 import { SurveyStatusLabelPipe } from '@pipes/survey-status-label.pipe';
 import { SurveyStatusSeverityPipe } from '@pipes/survey-status-severity.pipe';
 import { ConfirmationService } from 'primeng/api';
@@ -41,6 +42,8 @@ import { debounceTime, distinctUntilChanged, map, startWith } from 'rxjs';
     SurveyStatusSeverityPipe,
     DueDateLabelPipe,
     DueDateLabelColorPipe,
+    ScheduledSendTooltipPipe,
+    ScheduledSendAriaLabelPipe,
     TooltipModule,
     ConfirmDialogModule,
     EmptyStateComponent,
@@ -159,32 +162,6 @@ export class SurveysTableComponent {
     const responses = survey.total_responses || 0;
     const percentage = total > 0 ? Math.round((responses / total) * 100) : 0;
     return `${responses} of ${total} responses (${percentage}%)`;
-  }
-
-  protected getScheduledSendTooltip(survey: Survey): string {
-    if (!survey.survey_send_date) {
-      return '';
-    }
-    const sendDate = new Date(survey.survey_send_date);
-    if (Number.isNaN(sendDate.getTime())) {
-      return '';
-    }
-    const formatted = sendDate.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
-    return `Sends on ${formatted}`;
-  }
-
-  protected getScheduledStatusAriaLabel(survey: Survey, openResponded: boolean): string | null {
-    const tooltip = this.getScheduledSendTooltip(survey);
-    if (!tooltip) {
-      return null;
-    }
-    const statusLabel = openResponded ? 'Submitted' : this.getStatusLabel(survey);
-    return `${statusLabel}. ${tooltip}`;
-  }
-
-  private getStatusLabel(survey: Survey): string {
-    const displayStatus = getSurveyDisplayStatus(survey);
-    return SURVEY_STATUS_LABELS[displayStatus] ?? displayStatus;
   }
 
   // === Private Initializers ===
